requirements.txt
streamlit>=1.28.0
atlassian-python-api>=3.41.0
pandas>=2.0.0
requests>=2.31.0
reportlab>=4.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
